Added MediatR to HistoryController
