+++ Store.Web/Application/HistoryNote/Queries/GetHistoryNotesQuery.cs
@@ -1,0 +1,10 @@
+﻿using MediatR;
+using Store.Web.Dtos.HistoryNote;
+
+namespace Store.Web.Application.HistoryNote.Queries
+{
+    public class GetHistoryNotesQuery : IRequest<List<HistoryNoteHistoryViewDto>>
+    {
+
+    }
+}
