+++ Store.Web/Controllers/HistoryController.cs
@@ -1,41 +1,22 @@
-﻿using AutoMapper;
+﻿using MediatR;
 using Microsoft.AspNetCore.Mvc;
-using Store.Web.Abstractions.Data;
-using Store.Web.Dtos.HistoryNote;
-using Store.Web.Models;
+using Store.Web.Application.HistoryNote.Queries;
 
 namespace Store.Web.Controllers
 {
     public class HistoryController : Controller
     {
-        private readonly IHistoryNoteRepo<HistoryNote> _repo;
-        private readonly IMapper _mapper;
-        private readonly ILogger<HistoryController> _logger;
+        private readonly IMediator _mediator;
 
-        public HistoryController(IHistoryNoteRepo<HistoryNote> repo, 
-                                 IMapper mapper,
-                                 ILogger<HistoryController> logger)
+        public HistoryController(IMediator mediator)
         {
-            _repo = repo;
-            _mapper = mapper;
-            _logger = logger;
+            _mediator = mediator;
         }
 
-        public IActionResult HistoryMainView()
+        public async Task<IActionResult> HistoryMainView()
         {
-           
-            try
-            {
-                var notes = _repo.GetHistoryNotes().Result.ToList();
-                _logger.LogInformation("HistoryMainView: notes was successfully received");
-
-                return View(_mapper.Map<List<HistoryNote>, List<HistoryNoteHistoryViewDto>>(notes));
-            }
-            catch (Exception e)
-            {
-                _logger.LogError($"HistoryMainView: unable to get notes. Reason: {e.Message}");
-                return BadRequest();
-            }            
+            var notes = await _mediator.Send(new GetHistoryNotesQuery());
+            return View(notes);
         }
     }
 }
